Populate line data with real values
